+++ app/Models/Notification.php
@@ -215,6 +215,7 @@ class Notification extends Model
         return [
             'data' => 'array',
             'read_at' => 'datetime',
+            'push_wyslano_at' => 'datetime',
         ];
     }